feat: 完善API + 修复核心问题
- 新增 WebConfig: 注册 JwtInterceptor + CORS
- 新增 DashboardController: /api/dashboard/stats 统计接口
- 新增 SearchRequest + KnowledgeController POST /search
- 新增 GlobalExceptionHandler: 规范 HTTP 状态码 (404/400/401)
- 新增 DatabaseSchemaFix: 数据库字段调整
- 修复 auth/me: Integer userId 参数类型 + 拦截器注册
- 修复 CustomerRequest: 添加 industry 字段
- 修复 LeadRequest/LeadResponse/Lead: 添加 name/company/phone/email
- 修复 Deal: contractDate → dealDate (Request/Response/Entity)
- 修复 KnowledgeRequest/KnowledgeResponse: 添加 title/category
- 修复 FollowUp: nextFollowUpDate 字段
- 修复 UserService: Long → Integer ID 类型
- 添加商机阶段更新 + 线索转换 + 知识库 CRUD

+++ backend/src/main/java/com/crm/controller/AuthController.java
@@ -32,7 +32,7 @@ public class AuthController {
     }
     
     @GetMapping("/me")
-    public ResponseEntity<LoginResponse> getCurrentUser(@RequestAttribute Long userId) {
+    public ResponseEntity<LoginResponse> getCurrentUser(@RequestAttribute Integer userId) {
         LoginResponse response = userService.getUserInfo(userId);
         return ResponseEntity.ok(response);
     }
